[Refactor] Extract states.py from model_runner (#122)
