+++ vllm_metal/v1/states.py
@@ -1,0 +1,73 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Request state, sampling output, and small utilities for Metal model runner."""
+
+from dataclasses import dataclass, field
+from typing import TypeAlias
+
+import mlx.core as mx
+import torch
+from mlx_lm.models.cache import ArraysCache, KVCache, RotatingKVCache
+from vllm.sampling_params import SamplingParams
+
+# Type alias for any per-layer cache type supported by the model.
+#
+# Notes:
+# - Some models (e.g. gpt_oss) use `RotatingKVCache` for sliding-window attention.
+# - Hybrid models use `ArraysCache` for non-attention state.
+AnyCache: TypeAlias = KVCache | RotatingKVCache | ArraysCache
+
+
+def _mlx_greedy_sample(logits: mx.array) -> mx.array:
+    """Native MLX greedy sampling - avoids PyTorch round-trip.
+
+    Args:
+        logits: Logits tensor of shape (batch_size, vocab_size)
+
+    Returns:
+        Token IDs of shape (batch_size,)
+    """
+    return mx.argmax(logits, axis=-1)
+
+
+def _create_request_generator(
+    device: torch.device,
+    sampling_params: SamplingParams,
+) -> torch.Generator | None:
+    """Create a per-request generator for seeded sampling.
+
+    vLLM uses a per-request generator only when an explicit seed is provided.
+    For unseeded sampling, vLLM relies on the global RNG state.
+    """
+    if sampling_params.seed is None:
+        return None
+    if sampling_params.temperature < 1e-5:
+        return None
+    generator = torch.Generator(device=device)
+    generator.manual_seed(sampling_params.seed)
+    return generator
+
+
+@dataclass
+class SamplerOutput:
+    """Output from the sampler."""
+
+    token_ids: list[int]
+    logprobs: list[float] | None = None
+
+
+@dataclass
+class RequestState:
+    """State for an ongoing request with KV cache."""
+
+    token_ids: list[int]
+    # Length of the original prompt (prefix) within `token_ids`.
+    # vLLM applies repetition penalties to both prompt+output tokens, but applies
+    # presence/frequency penalties only to generated (output) tokens.
+    prompt_len: int
+    cache: list[AnyCache]  # Per-layer caches (KVCache, RotatingKVCache, or ArraysCache)
+    sampling_params: SamplingParams  # Sampling parameters for this request
+    generator: torch.Generator | None = None
+    generated_tokens: int = 0
+    block_ids: list[int] = field(
+        default_factory=list
+    )  # Scheduler-assigned paged KV blocks
